requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-migrate==4.0.5
flask-wtf==1.2.1
werkzeug==2.3.7
python-dotenv==1.0.0
email-validator==2.1.0
pyinstaller==6.5.0
pywebview==4.4.1